[ctr/video] Rewrite video pipeline to allow for 800x480 at 60hz
No longer supports Old 3DS. Buy a New 3DS.

Some of this code is hacky. It is intended to be used with 800x480 video
frames. Smaller and larger sizes have not been tested. There is no error
handling for frame sizes larger than the 1024x512 GPU buffers.

GPU Right loss is not properly handled. Do not press the home button.

* Adds perf counters on the bottom screen, shows decoding and blit times
* Removes the usage of Y2R, which was the primary bottleneck
  * MVD can output RGB565, and it also adds dithering
* Use PICA for video rotation, instead of horribly slow CPU-side
  rotation
  * The code was "prematurely optimized" which significantly impacted
    perf
  * PICA can do video rotation under a millisecond
  * Buffers are transfered between FCRAM and VRAM, since FCRAM accesses
    are painfully slow
  * Avoids pulling in Citro3D as a dependency as the command list setup
    is extremely simple